chore(package): add help and version commands
For #2
