+++ cmd/version.js
@@ -1,0 +1,5 @@
+const { version } = require('../package.json')
+
+module.exports = () => {
+  console.log(`v${version}`);
+}
